Task Manager
This program was coded in python. It is a simple task manager which allows users to log on and then view and edit tasks which are assigned to them. New tasks and users can be added to the program. .txt files are used to keep track of all of the users and tasks.